% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as_quaternion}
\alias{as_quaternion}
\title{Quaternion class}
\usage{
as_quaternion(x)
}
\arguments{
\item{x}{Object of class \code{"euler"}}
}
\description{
Build an unit real quaternion
}
\examples{
euler1 <- c(90, 0, 10)
euler1 dplyr::\%>\%
  to_euler() dplyr::\%>\%
  as_quaternion()
}
\seealso{
\code{\link[=to_euler]{to_euler()}} for class \code{"euler"}
}
