% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cartesian_comp}
\alias{cartesian_comp}
\alias{euler_cart2geo}
\alias{euler_geo2cart}
\title{Calculate the position and magnitude of the rotation pole from angular
velocity components.}
\usage{
euler_cart2geo(wx, wy, wz)

euler_geo2cart(lat, lon, mag)
}
\arguments{
\item{wx, wy, wz}{numeric. Cartesian velocity components (in any consistent
units).}

\item{lat, lon}{numeric.  Latitude and longitude of the Euler pole in degrees}

\item{mag}{numeric. Magnitude of rotation in any consistent units}
}
\value{
\code{\link[=euler_cart2geo]{euler_cart2geo()}} returns three element vector containing the
geographic position of the Euler
pole and the rotation magnitudes in the units given by the Cartesian
components.

\code{\link[=euler_geo2cart]{euler_geo2cart()}} returns the Cartesian components of
the rotation around the unit vectors.
}
\description{
Calculate the position and magnitude of the rotation pole from angular
velocity components.
}
\examples{
euler_cart2geo(wx = -0.001583, wy = 0.005064, wz = -0.010430) # should be: c(-63.036, 107.359, deg2rad(0.6705))
euler_geo2cart(-63.036, 107.359, deg2rad(0.6705))
}
