% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{common_greatcircle}
\alias{common_greatcircle}
\title{Great circle of Euler poles}
\usage{
common_greatcircle(x, y)
}
\arguments{
\item{x, y}{two-column vectors giving the geographic coordinates latitude
and longitude in degrees}
}
\description{
Calculates the pole to the great circle of Euler poles
}
