% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{common_smallcircle}
\alias{common_smallcircle}
\title{Common small circle of absolute and relative pole}
\usage{
common_smallcircle(r1, r2)
}
\arguments{
\item{r1, r2}{four-column vectors  giving the Cartesian coordinates of the
Euler vector and the amount of rotation in radians for first rotation
(\code{r1}) and subsequent second rotation (\code{r2})}
}
\description{
Common small circle of absolute and relative pole
}
\examples{
r1 <- c(90, 0, 0.7) \%>\% to_euler()
r2 <- c(45, 30, 0.15) \%>\% to_euler()
common_smallcircle(r1, r2)
}
