% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_factors}
\alias{count_factors}
\title{Count all factors of a number}
\usage{
count_factors(x)
}
\arguments{
\item{x}{Value to count factors of}
}
\description{
Count all factors of a number
}
\examples{
euler:::count_factors(1000)
}
