% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_11}
\alias{data_11}
\title{Matrix of integers for the Euler problem 11}
\format{A \code{matrix} consisting of 20 rows and 20 columns}
\source{
\url{https://projecteuler.net/problem=11}
}
\usage{
data_11
}
\description{
Matrix of integers for the Euler problem 11
}
\keyword{datasets}
