% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_13}
\alias{data_13}
\title{Matrix of integers for the Euler problem 13}
\format{A \code{matrix} consisting of 100 rows and 50 columns}
\source{
\url{https://projecteuler.net/problem=13}
}
\usage{
data_13
}
\description{
Matrix of integers for the Euler problem 13
}
\keyword{datasets}
