% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_18}
\alias{data_18}
\title{Lower triangle data for the Euler problem 18}
\format{A \code{matrix} consisting of 15 rows and 15 columns}
\source{
\url{https://projecteuler.net/problem=18}
}
\usage{
data_18
}
\description{
Lower triangle data for the Euler problem 18
}
\keyword{datasets}
