% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_67}
\alias{data_67}
\title{Lower triangle data for the Euler problem 67}
\format{A \code{matrix} consisting of 100 rows and 100 columns}
\source{
\url{https://projecteuler.net/project/resources/p067_triangle.txt}
}
\usage{
data_67
}
\description{
Lower triangle data for the Euler problem 67
}
\keyword{datasets}
