% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_8}
\alias{data_8}
\title{Vector of integers for the Euler problem 8}
\format{A \code{vector} consisting of 1000 \code{integer}s}
\source{
\url{https://projecteuler.net/problem=11}
}
\usage{
data_8
}
\description{
Vector of integers for the Euler problem 8
}
\keyword{datasets}
