% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{draw_card}
\alias{draw_card}
\title{Draw a card from chance or community chest}
\usage{
draw_card(deck)
}
\arguments{
\item{deck}{the deck to draw from}
}
\value{
A list with the drawn card and the new deck
}
\description{
Draw a card from chance or community chest
}
