% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{esieve}
\alias{esieve}
\title{Sieve of Eratosthenes for generating primes 2:n}
\source{
\url{https://en.wikipedia.org/wiki/Sieve_of_Eratosthenes}
}
\usage{
esieve(n)
}
\arguments{
\item{n}{An integer. Find all primes below this number.}
}
\value{
A vector of integers
}
\description{
Sieve of Eratosthenes for generating primes 2:n
}
\examples{
euler:::esieve(10)
}
