% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euclid}
\alias{euclid}
\title{Generate Pythagorean triples using Euclid's formula}
\source{
\url{https://en.wikipedia.org/wiki/Pythagorean_triple#Generating_a_triple}
}
\usage{
euclid(m, n, k = 1)
}
\arguments{
\item{m, n}{Positive integers where m > n > 0}

\item{k}{A multiplier}
}
\value{
A list containing a, b, and c
}
\description{
Generate Pythagorean triples using Euclid's formula
}
\details{
From the fact that for positive integers m and n, m > n, the a, b, and c
given by the formula are all positive integers
}
\examples{
euler:::euclid(2, 1)
}
