% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{euler}
\alias{euler}
\alias{to_euler}
\alias{from_euler}
\title{Euler class}
\usage{
to_euler(g)

from_euler(x)
}
\arguments{
\item{g}{three-column vector or 3*n matrix of the
geographic coordinates latitude and longitude, and the amount of rotation in
degrees}

\item{x}{Object of class \code{"euler"}, i.e. 4-column vector or 3*n matrix of the
Cartesian coordinates and the amount of rotation in radians}
}
\description{
Converts Euler pole in from geographic to Cartesian coordinates and Euler
angle from degrees to radians and back
}
\examples{
euler1 <- c(90, 0, 10)
to_euler(euler1)
to_euler(euler1) \%>\% from_euler()
}
