% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler67.R
\name{euler67}
\alias{euler67}
\title{Euler 67}
\usage{
euler67(triangle)
}
\arguments{
\item{triangle}{A lower triangle matrix}
}
\description{
This is a solution to 
\href{https://projecteuler.net/problem=67}{Euler problem 67}.
}
\examples{
euler67(triangle)

}
