% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_1.R
\name{euler_1}
\alias{euler_1}
\title{Project Euler 1: Multiples of 3 and 5}
\usage{
euler_1(x)
}
\arguments{
\item{x}{Max value}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=1}{Euler problem 1}.
}
\examples{
euler_1(1000)

}
