% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_10.R
\name{euler_10}
\alias{euler_10}
\title{Project Euler 10: Summation of primes}
\usage{
euler_10(n)
}
\arguments{
\item{n}{An integer to find the summation of primes under this value.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=10}{Euler problem 10}.
}
\examples{
euler_10(1000)

}
