% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_11.R
\name{euler_11}
\alias{euler_11}
\title{Project Euler 11: Largest product in a grid}
\usage{
euler_11(data)
}
\arguments{
\item{data}{A matrix of integers.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=11}{Euler problem 11}.
}
\examples{
euler_11(data_11)

}
