% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_12.R
\name{euler_12}
\alias{euler_12}
\title{Project Euler 12: Highly divisible triangular number}
\usage{
euler_12(x)
}
\arguments{
\item{x}{A number of interest to find the first triangle number with that
many divisors.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=12}{Euler problem 12}.
}
\examples{
euler_12(10)

}
