% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_13.R
\name{euler_13}
\alias{euler_13}
\title{Project Euler 13: Large sum}
\usage{
euler_13(data)
}
\arguments{
\item{data}{A matrix of integers.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=13}{Euler problem 13}.
}
\examples{
euler_13(data_13)

}
