% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_14.R
\name{euler_14}
\alias{euler_14}
\title{Project Euler 14: Longest Collatz sequence}
\usage{
euler_14(num)
}
\arguments{
\item{num}{A number for which to find the longest Collatz number up to.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=14}{Euler problem 14}.
}
\details{
Calculates the length of the chain for each Collatz number from 1 to the
number specified by \code{num}.
}
\examples{
euler_14(1e6)

}
