% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_67.R
\name{euler_18}
\alias{euler_18}
\alias{euler_67}
\title{Project Euler 18 (67): Maximum path sum I (II)}
\usage{
euler_18(data)

euler_67(data)
}
\arguments{
\item{data}{A lower triangle matrix.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=18}{Euler problem 18} and
\href{https://projecteuler.net/problem=67}{Euler problem 67}.
}
\details{
The code that solves both of these problems is the same.
}
\examples{
euler_18(data_18)
euler_67(data_67)

}
