% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_2.R
\name{euler_2}
\alias{euler_2}
\title{Project Euler 2: Even Fibonacci numbers}
\usage{
euler_2(x)
}
\arguments{
\item{x}{Max value}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=2}{Euler problem 2}.
}
\examples{
euler_2(4000000)

}
