% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_3.R
\name{euler_3}
\alias{euler_3}
\title{Project Euler 3: Largest prime factor}
\usage{
euler_3(x)
}
\arguments{
\item{x}{A number of interest to find the largest prime factor of.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=3}{Euler problem 3}.
}
\examples{
euler_3(600851475143)

}
