% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_4.R
\name{euler_4}
\alias{euler_4}
\title{Project Euler 4: Largest palindrome product}
\usage{
euler_4(x)
}
\arguments{
\item{x}{N digits}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=4}{Euler problem 4}.
}
\examples{
euler_4(2)

}
