% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_5.R
\name{euler_5}
\alias{euler_5}
\title{Project Euler 5: Smallest multiple}
\usage{
euler_5()
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=5}{Euler problem 5}.
}
\examples{
euler_5()

}
