% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_6.R
\name{euler_6}
\alias{euler_6}
\title{Project Euler 6: Sum square difference}
\usage{
euler_6(n)
}
\arguments{
\item{n}{An integer.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=6}{Euler problem 6}.
}
\examples{
euler_6(10)
euler_6(100)

}
