% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_67.R
\name{euler_67}
\alias{euler_67}
\title{Project Euler 67: Maximum path sum II}
\usage{
euler_67(data)
}
\arguments{
\item{data}{A lower triangle matrix.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=67}{Euler problem 67}.
}
\examples{
euler_67(data_67)

}
