% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_7.R
\name{euler_7}
\alias{euler_7}
\title{Project Euler 7: 10001st prime}
\usage{
euler_7(n)
}
\arguments{
\item{n}{An integer.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=7}{Euler problem 7}.
}
\examples{
euler_7(6)
euler_7(10001)

}
