% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_8.R
\name{euler_8}
\alias{euler_8}
\title{Project Euler 8: Largest product in a series}
\usage{
euler_8(data, n)
}
\arguments{
\item{data}{A vector of integers.}

\item{n}{An integer. The number of adjacent integers.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=8}{Euler problem 8}.
}
\examples{
euler_8(data_8, 4)
euler_8(data_8, 13)

}
