% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_9.R
\name{euler_9}
\alias{euler_9}
\title{Project Euler 9: Special Pythagorean triplet}
\usage{
euler_9(n)
}
\arguments{
\item{n}{An integer.}
}
\description{
This is a solution to
\href{https://projecteuler.net/problem=9}{Euler problem 9}.
}
\examples{
euler_9(3 + 4 + 5)
euler_9(1000)

}
