% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{euler_cart2geo}
\alias{euler_cart2geo}
\title{Calculate the position and magnitude of the rotation pole from angular
velocity components.}
\usage{
euler_cart2geo(w)
}
\arguments{
\item{w}{three-element vector containing the cartesian velocity components (rad or rad/myr).}
}
\value{
three element vector containing the geographic position of the Euler
pole and the rotation magnitudes in degrees.
}
\description{
Calculate the position and magnitude of the rotation pole from angular
velocity components.
}
\examples{
pa_nuvel <- c(x = -0.001583, y = 0.005064, z = -0.010430)
euler_cart2geo(pa_nuvel) # should be: c(-63.036, 107.359, 0.6705)
}
