% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaternion.R
\name{euler_concatenation}
\alias{euler_concatenation}
\title{Concatenation of rotations}
\usage{
euler_concatenation(r1, r2)
}
\arguments{
\item{r1, r2}{Objects of class \code{"euler"}, i.e. four-column vectors
giving the Cartesian coordinates of the
Euler vector and the amount of rotation in radians for first rotation
(\code{r1}) and subsequent second rotation (\code{r2})}
}
\value{
\code{list}. Euler axes (geographical coordinates) and Euler
angles (in degrees)
}
\description{
Accumulation of the two rotations r1 followed by r2
}
\examples{
x <- c(27.1275, 17.3248, 0.4024) \%>\% to_euler()
y <- c(22.2079, -92.4055, 0.0858) \%>\% to_euler()
euler_concatenation(x, y)
}
