% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greiner.R
\name{euler_matrix}
\alias{euler_matrix}
\title{Euler rotation matrix}
\usage{
euler_matrix(x)
}
\arguments{
\item{x}{Object of class \code{"euler"}}
}
\value{
\code{matrix}
}
\description{
Creates a matrix from the given set of values.
}
\examples{
ep <- c(0, 90, 90) \%>\% to_euler()
euler_matrix(ep)
}
\references{
Greiner, B. (1999). Euler rotations in plate-tectonic
reconstructions. \emph{Computers and Geosciences}, 25(3), 209--216.
\doi{10.1016/S0098-3004(98)00160-5}
}
