% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{euler_migration}
\alias{euler_migration}
\title{Euler migration path}
\usage{
euler_migration(x, y, steps = seq(1, 300, 25))
}
\arguments{
\item{x, y}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation
(\code{x}) and subsequent second rotation (\code{y})}

\item{steps}{numeric vector of time increments. The default is a sequence
from 1 to 10 by an incremental step of 1 (e.g. Myr)}
}
\description{
Euler migration path
}
\examples{
x <- c(90, 0, 0.7)
y <- c(45, 30, 0.15)
euler_migration(x, y)
}
