% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fibs}
\alias{fibs}
\title{fibs
get fibonacci numbers}
\usage{
fibs(n = 10, x1 = 0, x2 = 1)
}
\arguments{
\item{n}{the index of the fibonacci number ot return (default: 10)}

\item{x1}{the first number in the sequence (default: 0)}

\item{x2}{the second number in the sequence (default: 1)}
}
\value{
numeric: the nth fibonacci number
}
\description{
fibs
get fibonacci numbers
}
