% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_factors}
\alias{find_factors}
\title{Find all factors of a number}
\usage{
find_factors(x)
}
\arguments{
\item{x}{Value to find factors of}
}
\value{
List of vectors of factor pairs
}
\description{
Find all factors of a number
}
\examples{
euler:::find_factors(1000)
}
