% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{finite_euler}
\alias{finite_euler}
\title{Finite Euler rotation}
\usage{
finite_euler(r1, r2)
}
\arguments{
\item{r1}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{r1})
and subsequent second rotation (\code{r2})}

\item{r2}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{r1})
and subsequent second rotation (\code{r2})}
}
\description{
Finite Euler rotation
}
