% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gc_distance}
\alias{gc_distance}
\title{Great circle distance}
\usage{
gc_distance(a, b)
}
\arguments{
\item{a, b}{two-column vectors giving the geographic coordinates latitude
and longitude in degrees}
}
\description{
Angle between two points on a sphere, measured along great circle passing
through both point vectors
}
