% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_circular_permutations}
\alias{get_all_circular_permutations}
\title{get_all_circular_permutations}
\usage{
get_all_circular_permutations(number)
}
\arguments{
\item{number}{the number}
}
\description{
get all of the circular permutations of a number
}
