% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_digits}
\alias{get_digits}
\title{get_digits}
\usage{
get_digits(number)
}
\arguments{
\item{number}{the number to get permutations of}
}
\value{
the digits
}
\description{
get the digits of a number
}
