% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_factors}
\alias{get_factors}
\title{get_factors
get the factors of a number}
\usage{
get_factors(n)
}
\arguments{
\item{n}{the number}
}
\value{
numeric: the factors
}
\description{
get_factors
get the factors of a number
}
