% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_nth_triangle}
\alias{get_nth_triangle}
\title{get_nth_triangle}
\usage{
get_nth_triangle(n)
}
\arguments{
\item{n}{the nth number}
}
\value{
the nth triangle number
}
\description{
get the nth triangle number
}
