% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_number_digit_permutation}
\alias{get_number_digit_permutation}
\title{get_number_digit_permutation}
\usage{
get_number_digit_permutation(n, digits)
}
\arguments{
\item{n}{number of digits to move}

\item{digits}{the digits of the number to permute}
}
\value{
the number
}
\description{
get all circular variants of a number
}
