% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_squares}
\alias{get_squares}
\title{get_squares}
\usage{
get_squares(n_max = 10, type = "all")
}
\arguments{
\item{n_max}{the number of odd squares to get}

\item{type}{'odd' for the odd squares, 'even' for the even squares, 'all' for all squares}
}
\value{
the numbers
}
\description{
get squares (either even, odd, or all)
}
