% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_triangle_numbers}
\alias{get_triangle_numbers}
\title{get_triangle_numbers}
\usage{
get_triangle_numbers(n)
}
\arguments{
\item{n}{the max n}
}
\description{
get all of the triangle numbers up to t(n)
}
