% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_word_values}
\alias{get_word_values}
\title{get_word_values}
\usage{
get_word_values(word)
}
\arguments{
\item{word}{the letter to convert}
}
\value{
the number
}
\description{
convert a word to a number based on the num of its letters
}
