% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{init_community_chest}
\alias{init_community_chest}
\title{Initialize community chest deck}
\usage{
init_community_chest()
}
\value{
a character vector of randomly shuffled community chest cards
}
\description{
Initialize community chest deck
}
