% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inverse_euler}
\alias{inverse_euler}
\title{Inverse rotation}
\usage{
inverse_euler(x)
}
\arguments{
\item{x}{Object of class \code{"euler"}, i.e. 4-column vector or 3*n matrix of the
Cartesian coordinates and the amount of rotation in radians}
}
\description{
Inverse rotation
}
