% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_palindrome}
\alias{is_palindrome}
\title{Check if a number is a palindrome, i.e. it remains the same
whether read from left to right or right to left.}
\usage{
is_palindrome(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
TRUE if it is a palindrome, FALSE otherwise
}
\description{
Check if a number is a palindrome, i.e. it remains the same
whether read from left to right or right to left.
}
\examples{
euler:::is_palindrome(1001)
}
