% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_prime}
\alias{is_prime}
\title{is_prime
check if a number is prime}
\usage{
is_prime(n)
}
\arguments{
\item{n}{the number}
}
\value{
boolean: TRUE if prime, FALSE otherwise
}
\description{
is_prime
check if a number is prime
}
