% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lepichon}
\alias{lepichon}
\alias{as_quaternion2}
\alias{quat_composition}
\alias{quat_2_angles}
\title{Le Pichon method}
\usage{
as_quaternion2(x)

quat_composition(q1, q2)

quat_2_angles(q)
}
\description{
Le Pichon method
}
