% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{letter_to_number}
\alias{letter_to_number}
\title{letter_to_number}
\usage{
letter_to_number(letter)
}
\arguments{
\item{letter}{the letter to convert}
}
\value{
the number
}
\description{
convert a letter to a number
}
