% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greiner.R
\name{matrix_2_angles}
\alias{matrix_2_angles}
\title{Euler axis and angle from Euler matrix}
\usage{
matrix_2_angles(A)
}
\arguments{
\item{A}{3x3 matrix}
}
\value{
axis and angle in degrees
}
\description{
Extracts the coordinates of an Euler pole and the angle from a
Euler matrix
}
\details{
If there is no rotation (i.e. angle = 0), the coordinates of the
axis are equal to Earth's spin axis (according to the GPLATES convention).
}
\examples{
c(90, 0, 90) \%>\%
  to_euler() \%>\%
  euler_matrix() \%>\%
  matrix_2_angles()
}
