% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{monopoly_roll}
\alias{monopoly_roll}
\title{Monopoply Roll}
\usage{
monopoly_roll(sides = 4)
}
\arguments{
\item{sides}{the number of sides on the dice}
}
\value{
a numeric vector of the die rolls
}
\description{
Rolling two dice
}
