% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{next_railroad}
\alias{next_railroad}
\title{Go to the next railroad}
\usage{
next_railroad(current_square)
}
\arguments{
\item{current_square}{the index of the current square}
}
\value{
The square of the next railroad
}
\description{
Go to the next railroad
}
