% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{next_utility}
\alias{next_utility}
\title{Go to the next utility}
\usage{
next_utility(current_square)
}
\arguments{
\item{current_square}{the index of the current square}
}
\value{
The square of the next utility
}
\description{
Go to the next utility
}
