% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_vector}
\alias{normalize_vector}
\title{Normalization of a vector}
\usage{
normalize_vector(v)
}
\arguments{
\item{v}{Numeric vector}
}
\description{
normalizes a vector to unit length (length = 1)
}
