% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{plate_rotation}
\alias{plate_rotation}
\title{Rotate plate using a set of migrated poles}
\usage{
plate_rotation(x, p)
}
\arguments{
\item{x}{\code{data.frame} with time, lat, lon, and angle}

\item{p}{\code{sf} object. Plate with location at 0 Ma}
}
\description{
Rotate plate using a set of migrated poles
}
