% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{pole_migration}
\alias{pole_migration}
\title{Euler pole migration}
\usage{
pole_migration(
  x,
  y,
  steps = c(1, seq(25, 300, 25)),
  infinitesimal = TRUE,
  finite = TRUE
)
}
\arguments{
\item{x}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{x})
and subsequent second rotation (\code{y})}

\item{y}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{x})
and subsequent second rotation (\code{y})}

\item{steps}{numeric vector of time increments. The default is a sequence
from 1 to 10 by an incremental step of 1 (e.g. Myr)}

\item{infinitesimal}{logical. Should the rotation be calculated using
the infinitesimal and/or the finite rotation approach?}

\item{finite}{logical. Should the rotation be calculated using
the infinitesimal and/or the finite rotation approach?}
}
\description{
Migration of the relative Euler pole associated with two absolute plate motions
}
\examples{
in.eu <- c(27.12746847, 17.32482497, 0.402388191)
som.eu <- c(22.2078593, -92.40545103, 0.085835298)
pole_migration(som.eu, in.eu)
pole_migration(in.eu, som.eu)
}
