% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{pole_migration_stats}
\alias{pole_migration_stats}
\title{Euler pole migration statistics}
\usage{
pole_migration_stats(x, euler1, euler2)
}
\arguments{
\item{x}{\code{data.frame}. Output from \code{pole_migration}}

\item{euler1, euler2}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{x})
and subsequent second rotation (\code{y})}
}
\value{
two-column vector containing magnitude of the vector between the
Euler poles \eqn{||d||} (\code{d}),
... in degree \eqn{\eta} (\code{eta}),
and the change of their great circle distance in degree \eqn{\Delta} (\code{delta}).
}
\description{
Rates of the migration of the relative Euler pole associated with two absolute plate motions
}
\examples{
in.eu <- c(27.12746847, 17.32482497, 0.402388191)
som.eu <- c(22.2078593, -92.40545103, 0.085835298)
pm <- pole_migration(som.eu, in.eu)
pole_migration_stats(pm, som.eu, in.eu)
}
\seealso{
\code{\link[=pole_migration]{pole_migration()}} for calculating the migration of the relative Euler pole.
}
