% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prime_factors}
\alias{prime_factors}
\title{prime_factors
get the prime factors of a number}
\usage{
prime_factors(n)
}
\arguments{
\item{n}{the number}
}
\value{
numeric: the prime factors
}
\description{
prime_factors
get the prime factors of a number
}
