% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{problem_1}
\alias{problem_1}
\title{Problem 1
If we list all the natural numbers below 10 that are multiples of 3 or 5, we get 3, 5, 6 and 9. The sum of these multiples is 23.
Find the sum of all the multiples of 3 or 5 below 1000.}
\usage{
problem_1()
}
\value{
the sum
}
\description{
Problem 1
If we list all the natural numbers below 10 that are multiples of 3 or 5, we get 3, 5, 6 and 9. The sum of these multiples is 23.
Find the sum of all the multiples of 3 or 5 below 1000.
}
