% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{problem_22}
\alias{problem_22}
\title{Problem 22}
\usage{
problem_22()
}
\value{
the total name score
}
\description{
Using names.txt (right click and 'Save Link/Target As...'), a 46K text file containing over five-thousand first names, begin by sorting it into alphabetical order. Then working out the alphabetical value for each name, multiply this value by its alphabetical position in the list to obtain a name score.
For example, when the list is sorted into alphabetical order, COLIN, which is worth 3 + 15 + 12 + 9 + 14 = 53, is the 938th name in the list. So, COLIN would obtain a score of 938 × 53 = 49714.
What is the total of all the name scores in the file?
}
