% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{problem_28}
\alias{problem_28}
\title{Problem 28}
\usage{
problem_28()
}
\value{
the sum of the numbers on the diagonals
}
\description{
Starting with the number 1 and moving to the right in a clockwise direction a 5 by 5 spiral is formed as follows:
}
\details{
21 22 23 24 25
20  7  8  9 10
19  6  1  2  3
18  5  4  3 12
17 16 15 14 13
It can be verified that the sum of the numbers on the diagonals is 101.
What is the sum of the numbers on the diagonals in a 1001 by 1001 spiral formed the same way?
}
