% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{problem_3}
\alias{problem_3}
\title{Problem 3
The prime factors of 13195 are 5, 7, 13 and 29.
What is the largest prime factor of the number 600851475143 ?}
\usage{
problem_3()
}
\value{
the largest prime factor
}
\description{
Problem 3
The prime factors of 13195 are 5, 7, 13 and 29.
What is the largest prime factor of the number 600851475143 ?
}
