% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{problem_35}
\alias{problem_35}
\title{Problem 35}
\usage{
problem_35(workers = 8)
}
\arguments{
\item{workers}{the number of workers to use for parallel computing}
}
\value{
the number of circular primes
}
\description{
The number, 197, is called a circular prime because all rotations of the digits: 197, 971, and 719, are themselves prime.
There are thirteen such primes below 100: 2, 3, 5, 7, 11, 13, 17, 31, 37, 71, 73, 79, and 97.
How many circular primes are there below one million?
}
