% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{problem_42}
\alias{problem_42}
\title{Problem 42}
\usage{
problem_42()
}
\value{
the number of triangle words
}
\description{
The nth term of the sequence of triangle numbers is given by, tn = ½n(n+1); so the first ten triangle numbers are:
1, 3, 6, 10, 15, 21, 28, 36, 45, 55, ...
By converting each letter in a word to a number corresponding to its alphabetical position and adding these values we form a word value. For example, the word value for SKY is 19 + 11 + 25 = 55 = t10. If the word value is a triangle number then we shall call the word a triangle word.
}
\details{
Using words.txt (right click and 'Save Link/Target As...'), a 16K text file containing nearly two-thousand common English words, how many are triangle words?
}
