% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{problem_84}
\alias{problem_84}
\title{Problem 84}
\usage{
problem_84(num_sims = 10000, num_die_sides = 4)
}
\arguments{
\item{num_sims}{The number of simulations to run}

\item{num_die_sides}{the number of sides of the monopoly dice}
}
\value{
the requested string
}
\description{
Problem 84
}
