% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solutions.R
\name{problem_84_roughly_exact}
\alias{problem_84_roughly_exact}
\title{A solution to 84 by approximating the transition matrix and directly solving from it}
\usage{
problem_84_roughly_exact()
}
\value{
the three most common squares
}
\description{
A solution to 84 by approximating the transition matrix and directly solving from it
}
