% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{quick_analysis}
\alias{quick_analysis}
\title{Quick analysis}
\usage{
quick_analysis(model = c("GSRM", "MORVEL"), plateA, plateB, fix)
}
\arguments{
\item{plateA, plateB}{plates to extract}

\item{fix}{Reference system that is considered to be fixed.}
}
\value{
list containing a plot, a gt table and data.frame with the results
}
\description{
Analysis the data and returns a ggplot and a gt table
}
\examples{
quick_analysis("GSRM", "IN", "SO", "EU")
}
