% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{relative_euler}
\alias{relative_euler}
\title{As-if-infinitesimal rotation}
\usage{
relative_euler(p, q)
}
\arguments{
\item{p}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{p})
and subsequent second rotation (\code{q})}

\item{q}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{p})
and subsequent second rotation (\code{q})}
}
\description{
As-if-infinitesimal rotation
}
