% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{relative_quat}
\alias{relative_quat}
\title{Infinitesimal rotation}
\usage{
relative_quat(p, q)
}
\arguments{
\item{p, q}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{p})
and subsequent second rotation (\code{q})}
}
\description{
Infinitesimal rotation using quaternions
}
