% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{relative_rotation}
\alias{relative_rotation}
\title{Relative Rotation associated to two absolute rotations}
\usage{
relative_rotation(x, y, infinitesimal = TRUE, finite = TRUE)
}
\arguments{
\item{x, y}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{x})
and subsequent second rotation (\code{y})}

\item{infinitesimal, finite}{logical. Should the rotation be calculated using
the infinitesimal and/or the finite rotation approach?}
}
\value{
\code{list}. Infinitesimal and the finite approach Euler axes
(geographical coordinates) and Euler angles (in degrees)
}
\description{
Calculates Euler rotation axis and angle for two given absolute rotations using the
infinitesimal (Schaeben et al., 2021) and the finite approach (Cox \& Hart, 1989)
}
\note{
\code{x} will be considered as the fixed plate for the relative plate
motion between \code{x} and \code{y}
}
\examples{
x <- c(90, 0, 0.7)
y <- c(45, 30, 0.15)
relative_rotation(x, y)
}
\references{
Schaeben, H., Kroner, U. and Stephan, T. (2021). Euler Poles of Tectonic
Plates. In B. S. Daza Sagar, Q. Cheng, J. McKinley and F. Agterberg (Eds.),
\emph{Encyclopedia of Mathematical Geosciences. Encyclopedia of Earth Sciences Series}
(pp. 1--7). Springer Nature Switzerland AG 2021.
}
\seealso{
\code{\link[=quasi_infinitesimal_euler]{quasi_infinitesimal_euler()}} and \code{\link[=relative_euler_schaeben]{relative_euler_schaeben()}} for quasi-infinitesimal and infinitesimal rotation, respectively.
}
