% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rotate_euler}
\alias{rotate_euler}
\title{Infinitesimal and finite Rotations}
\usage{
rotate_euler(x, y)
}
\arguments{
\item{x, y}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for rotation 1 \code{x}
and subsequent rotation 2}
}
\description{
get both infinitesimal and finite Euler rotation axis and angles
}
\examples{
x <- c(90, 0, 0.7)
y <- c(45, 30, 0.15)
rotate_euler(x, y)
}
