% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rotate_vector}
\alias{rotate_vector}
\title{Rotate a set a vector}
\usage{
rotate_vector(x, p)
}
\arguments{
\item{x}{Vector for Euler pole position, three-column vector of the
geographic coordinates latitude and longitude, and the amount of rotation in
degrees}

\item{p}{\code{sf} object}
}
\value{
\code{sf} object
}
\description{
Rotate a vector, set of points, or polygon by a given rotation
}
\examples{
plates <- readRDS("data/plates.rds")
india <- subset(plates, Code == "IN")
euler <- c(90, 0, 90)
rotate_vector(euler, india) \%>\% plot()
}
