% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greiner.R
\name{rotation_axis}
\alias{rotation_axis}
\title{Rotation axis from rotation matrix}
\usage{
rotation_axis(A, psi)
}
\arguments{
\item{A}{3x3 matrix}

\item{psi}{angle in radians}
}
\value{
Euler vector in Cartesian coordinates
}
\description{
Extracts the rotation axis from rotation matrix
}
\examples{
w <- c(0, 90, 90) \%>\% to_euler()
rot <- rotation_matrix(w)

rotation_axis(rot, rotation_angle(rot))
}
\seealso{
\code{\link[=rotation_angle]{rotation_angle()}} to extract the angle of the rotation matrix.
}
