% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greiner.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Rotation Matrix}
\usage{
rotation_matrix(x)
}
\arguments{
\item{x}{Object of class \code{"euler"}}
}
\value{
\code{matrix}
}
\description{
Calculates the rotation matrix using the rotation axis and the
angle of rotation
}
\examples{
w <- c(0, 90, 90) \%>\% to_euler()
rotation_matrix(w)
}
\seealso{
\code{\link[=rotation_axis]{rotation_axis()}} and \code{\link[=rotation_angle]{rotation_angle()}} to extract the axis and the
angle from the rotation matrix.
}
