% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{runShiny}
\alias{runShiny}
\title{Run the shiny application}
\usage{
runShiny(doDebug = FALSE, ...)
}
\arguments{
\item{doDebug}{boolean, whether to show debug options in the UI; default FALSE}

\item{...}{further arguments that can be passed to \code{\link[shiny]{shinyApp}} options}
}
\value{
shinyApp object, see also \code{\link[shiny]{shinyApp}}
}
\description{
Run the shiny application
}
