% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_main.R
\name{serverFunction}
\alias{serverFunction}
\title{Main server function (like in server.R)}
\usage{
serverFunction()
}
\value{
function that can be used as server component of a shiny app
}
\description{
Main server function (like in server.R)
}
