% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sf_conversion}
\alias{sf_conversion}
\alias{sf_to_vector}
\alias{vector_to_sf}
\title{Convert between sf objects and point vectors}
\usage{
sf_to_vector(sf)

vector_to_sf(x, multi = FALSE)
}
\arguments{
\item{sf}{\code{sf} object}

\item{x}{numeric vector}

\item{multi}{logical. Is x MULTIPOLYGON or MULTILINESTRING.}
}
\description{
Converts a sf object into a two-column vector or vice versa
}
\examples{
readRDS("data/plates.rds")
in.plate <- subset(plates, Code == "IN")
sf_to_vector(in.plate) \%>\% vector_to_sf()
}
