% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sf_to_vector}
\alias{sf_to_vector}
\title{sf object to vector}
\usage{
sf_to_vector(x)
}
\arguments{
\item{x}{\code{sf} object}
}
\description{
Converts a sf object into a two-column vector
}
\examples{
readRDS("data/plates.rds")
in.plate <- subset(plates, Code == "IN")
sf_to_vector(in.plate)
}
