% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{to_euler}
\alias{to_euler}
\title{Euler class}
\usage{
to_euler(x)
}
\arguments{
\item{x}{three-column vector or 3*n matrix of the
geographic coordinates latitude and longitude, and the amount of rotation in
degrees}
}
\description{
Converts Euler pole in from geographic to Cartesian coordinates and Euler
angle from degrees to radians
}
\examples{
euler1 <- c(90, 0, 10)
to_euler(euler1)
}
