% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{triangle}
\alias{triangle}
\title{Find triangle numbers}
\usage{
triangle(x)
}
\arguments{
\item{x}{How many triangle numbers to find}
}
\description{
Find triangle numbers
}
\examples{
euler:::triangle(10)
}
