% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{triangle_greater_than_max}
\alias{triangle_greater_than_max}
\title{triangle_greater_than_max}
\usage{
triangle_greater_than_max(n, max_num)
}
\arguments{
\item{n}{the number to test (i.e. the 10th triangle number)}

\item{max_num}{the max to test against}
}
\value{
a boolean (true if the nth triangle number is bigger than the max to test)
}
\description{
test if a triangle number is greater than some specified maximum value
}
