% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_main.R
\name{uiFunction}
\alias{uiFunction}
\title{Main UI function (like in ui.R)}
\usage{
uiFunction(doDebug)
}
\arguments{
\item{doDebug}{boolean, whether to show extra debug options in the UI}
}
\value{
UI object
}
\description{
Main UI function (like in ui.R)
}
