% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ut}
\alias{ut}
\title{ut
A function to simplify adding dependencies}
\usage{
ut(pack)
}
\arguments{
\item{pack}{the package to add to the description}
}
\value{
NULL (invisible)
}
\description{
ut
A function to simplify adding dependencies
}
