% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vector_to_sf}
\alias{vector_to_sf}
\title{vector to sf object}
\usage{
vector_to_sf(x, multi = FALSE)
}
\arguments{
\item{x}{vector}

\item{multi}{logical. Is x a MULTIPOLYGON or MULTILINESTRING.}
}
\description{
Converts a four-column vector into a sf object into
}
\examples{
readRDS("data/plates.rds")
in.plate <- subset(plates, Code == "IN")
sf_to_vector(in.plate) dplyr::\%>\% vector_to_sf()
}
