% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{euler}
\alias{euler}
\title{Euler class}
\usage{
euler(x)
}
\arguments{
\item{x}{Vector for Euler pole position, three-column vector of the
geographic coordinates latitude and longitude, and the amount of rotation in
degrees}
}
\description{
Converts Euler pole in from geographic to Cartesian coordinates and Euler
angle from degrees to radians
}
\examples{
euler1 <- euler(c(90, 0, 10))
euler2 <- euler(c(45, 30, 20))
}
