% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{euler_migration}
\alias{euler_migration}
\title{Euler pole migration}
\usage{
euler_migration(x, y, steps = c(1, seq(25, 300, 25)))
}
\arguments{
\item{x, y}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation
(\code{x}) and subsequent second rotation (\code{y})}

\item{steps}{numeric vector of time increments. The default is a sequence
from 1 to 10 by an incremental step of 1 (e.g. Myr)}
}
\description{
Migration of the relative Euler pole associated with two absolute plate motions
}
\examples{
in.eu <- c(27.12746847, 17.32482497, 0.402388191)
som.eu <- c(22.2078593, -92.40545103, 0.085835298)
euler_migration(in.eu, som.eu)
}
