% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{multiples_sum}
\alias{multiples_sum}
\title{multiples_sum
get the sum of the multiples of the numbers specified less than the max value specified}
\usage{
multiples_sum(numbers = c(3, 5), max_value = 1000)
}
\arguments{
\item{numbers}{the numbers whose multiples to include in the sum}

\item{max_value}{the largest number to consider}
}
\value{
the sum
}
\description{
multiples_sum
get the sum of the multiples of the numbers specified less than the max value specified
}
