% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{quick_analysis}
\alias{quick_analysis}
\title{Quick analysis}
\usage{
quick_analysis(model = c("GSRM", "MORVEL"), plateA, plateB, fix)
}
\arguments{
\item{model}{Model to choose from. Either \code{"GSRM"} for the "Global
Strain Rate Model" v2.1 by Kreemer et all. 2014, or \code{"MORVEL"} for the
"NNR-MORVEL56" model by Argus et al. 2011}

\item{plateA}{plates to extract}

\item{plateB}{plates to extract}

\item{fix}{Reference system that is considered to be fixed.}
}
\value{
list containing a plot, a gt table and \code{data.frame} with the results
}
\description{
Analysis the data and returns a ggplot and a gt table
}
\examples{
\dontrun{
quick_analysis("GSRM", "IN", "SO", "EU")
}
}
