% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{relative_rotation}
\alias{relative_rotation}
\title{Relative Rotation associated to two absolute rotations}
\usage{
relative_rotation(x, y, infinitesimal = TRUE, finite = TRUE)
}
\arguments{
\item{x, y}{three-column vectors  giving the geographic coordinates latitude
and longitude, and the amount of rotation in degrees for first rotation (\code{x})
and subsequent second rotation (\code{y})}

\item{infinitesimal, finite}{logical. Should the rotation be calculated using
the infinitesimal and/or the finite rotation approach?}
}
\value{
\code{list}. Infinitesimal and the finite approach Euler axes
(geographical coordinates) and Euler angles (in degrees)
}
\description{
Calculates Euler rotation axis and angle for two given absolute rotations using the
infinitesimal (Schaeben et al. 2021) and the finite approach
}
\note{
\code{x} will be considered as the fixed plate for the relative plate
motion between \code{x} and \code{y}
}
\examples{
x <- c(90, 0, 0.7)
y <- c(45, 30, 0.15)
relative_rotation(x, y)
}
\references{
Schaeben, H., Kroner, U., Stephan, T. (2021). Euler Poles
of Tectonic Plates. In B. S. Daza Sagar, Q. Cheng, J. McKinley,; F. Agterberg
(Eds.), Encyclopedia of Mathematical Geosciences. Encyclopedia of Earth
Sciences Series (pp. 1--7). Springer Nature Switzerland AG 2021.
\doi{10.1007/978-3-030-26050-7_435-1}
}
