% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{rotate_vector}
\alias{rotate_vector}
\title{Rotate a set a vector}
\usage{
rotate_vector(x, p, ...)
}
\arguments{
\item{x}{Object of class \code{"euler"}, i.e. 4-column vector or 3*n matrix of the
Cartesian coordinates and the amount of rotation in radians}

\item{p}{\code{sf} object}

\item{...}{additional arguments to daughter functions}
}
\value{
\code{sf} object
}
\description{
Rotate a vector, set of points, or polygon by a given rotation
}
\examples{
plates <- readRDS("data/plates.rds")
india <- subset(plates, Code == "IN")
euler <- c(90, 0, 90)
rotate_vector(euler, india) \%>\% plot()
}
