% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation}
\alias{rotation}
\alias{as_if_infinitesimal_euler}
\alias{finite_euler}
\alias{infinitesimal_euler}
\alias{infinitesimal_quaternion}
\title{Rotations}
\usage{
as_if_infinitesimal_euler(r1, r2)

finite_euler(r1, r2)

infinitesimal_euler(r1, r2)

infinitesimal_quaternion(r1, r2)
}
\arguments{
\item{r1, r2}{four-column vectors  giving the Cartesian coordinates of the
Euler vector and the amount of rotation in radians for first rotation
(\code{r1}) and subsequent second rotation (\code{r2})}
}
\description{
Rotations
}
\examples{
x <- c(90, 0, 0.7) \%>\% to_euler()
y <- c(45, 30, 0.15) \%>\% to_euler()
as_if_infinitesimal_euler(x, y)
finite_euler(x, y)
infinitesimal_euler2(x, y)
infinitesimal_euler(x, y)
}
