% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{triangle}
\alias{triangle}
\title{Lower triangle data for Euler problem 67}
\format{A \code{data.frame} consisting of 100 rows and 100 columns}
\source{
\url{https://projecteuler.net/problem=67}
}
\usage{
triangle
}
\description{
Lower triangle data for Euler problem 67
}
\keyword{datasets}
